process window works and opens on command. class structure is messy though.
