+++ Main.cpp
@@ -8,12 +8,12 @@ int main(int, char**)
     //bool wait = true;
     //Process valheimProc("valheim.x86_64", wait);
     //valheimProc.GetModuleBase("libmonobdwgc-2.0.so");
-    System mySystem;
-    //Overlay overlay;
-    // if (overlay.isInitialized)
-    // {
-    //     overlay.Register((Overlay::UserOLFunc)ConsoleWindow::MemPenguinOL);
-    //     overlay.Run();
-    // }
+    MainInterface mainUI;
+    Overlay overlay;
+    if (overlay.isInitialized)
+    {
+        overlay.Register((Overlay::UserOLFunc)mainUI.Update);
+        overlay.Run();
+    }
     return 0;
 }
