+++ include/Interface.h
@@ -1,26 +1,22 @@
 #pragma once
 
 #include "imgui.h"
+#include "System.h"
 
-class ConsoleWindow
+class ConsoleWindow;
+class ProcessWindow;
+
+class MainInterface
 {
 private:
-    char inputBuff[256];
-    ImVector<char*> items;
-    ImVector<const char*> commands;
-    ImVector<char*> history;
-    int historyPos;    // -1: new line, 0..history.Size-1 browsing history.
-    ImGuiTextFilter filter;
-    bool autoScroll;
-    bool scrollToBottom;
+    /* data */
 public:
-    ConsoleWindow();
-    ~ConsoleWindow();
-    void ClearLog();
-    void AddLog(const char* fmt, ...) IM_FMTARGS(2);
-    void Draw(const char* title, bool* p_open);
-    void ExecCommand(const char* command_line);
-    static int TextEditCallbackStub(ImGuiInputTextCallbackData* data);
-    static void MemPenguinOL();
-    int TextEditCallback(ImGuiInputTextCallbackData* data);
+    static bool showConsole;
+    static bool showProcWindow;
+    static ConsoleWindow console;
+    static ProcessWindow procWindow;
+    MainInterface();
+    ~MainInterface();
+    static void Update();
+    static MainInterface mainUI;
 };
